+++ LINBIT/PluginHelper.pm
@@ -2,8 +2,26 @@ package LINBIT::PluginHelper;
 
 use strict;
 use warnings;
+use Exporter 'import';
+our @EXPORT_OK = qw(valid_legacy_name valid_uuid_name valid_snap_name valid_name get_images);
 
 # use Data::Dumper;
+
+sub valid_legacy_name {
+    $_[0] =~ /^vm-\d+-disk-\d+\z/
+}
+
+sub valid_uuid_name {
+    $_[0] =~ /^pm-[\da-f]{8}_\d+\z/
+}
+
+sub valid_snap_name {
+    $_[0] =~ /^snap_.+_.+\z/
+}
+
+sub valid_name {
+    valid_legacy_name $_[0] or valid_uuid_name $_[0]
+}
 
 sub get_images {
     my ( $storeid, $vmid, $vollist, $resources, $node_name, $rg,
@@ -11,17 +29,29 @@ sub get_images {
       = @_;
 
     my $res = [];
-    foreach my $name ( keys %$resources ) {
+    foreach my $linstor_name ( keys %$resources ) {
 
         # skip if not on this node
-        next unless exists $resources->{$name}->{$node_name};
+        next unless exists $resources->{$linstor_name}->{$node_name};
 
         # skip if not from this RG
-        next unless $rg eq $resource_definitions->{$name}->{rg_name};
+        next unless $rg eq $resource_definitions->{$linstor_name}->{rg_name};
 
-        next unless $name =~ /^vm-(\d+)-/;
-        my $owner = $1;                 # aka "vmid"
-        my $volid = "$storeid:$name";
+        my $owner;
+        my $proxmox_name;
+        if ($linstor_name =~ /^pm-[\da-f]{8}\z/) {
+            $owner = $resource_definitions->{$linstor_name}->{vmid};
+            $proxmox_name = $linstor_name . "_" . $owner;
+            next unless valid_uuid_name($proxmox_name);
+        }
+        elsif ($linstor_name =~ /^vm-(\d+)-/) {
+            $owner = $1;    # aka "vmid"
+            $proxmox_name = $linstor_name;
+        } else {
+            next;
+        }
+
+        my $volid = "$storeid:$proxmox_name";
 
         # filter, if we have been passed vmid or vollist
         next if defined $vmid and $vmid ne $owner;
@@ -32,10 +62,10 @@ sub get_images {
         # expect exactly one volume
         # XXX warn for 0 or >= 2 volume resources?
         next
-          unless exists $resources->{$name}->{$node_name}->{nr_vols}
-          and $resources->{$name}->{$node_name}->{nr_vols} == 1;
+          unless exists $resources->{$linstor_name}->{$node_name}->{nr_vols}
+          and $resources->{$linstor_name}->{$node_name}->{nr_vols} == 1;
 
-        my $size_kib = $resources->{$name}->{$node_name}->{usable_size_kib};
+        my $size_kib = $resources->{$linstor_name}->{$node_name}->{usable_size_kib};
 
         push @$res,
           {