+++ LINSTORPlugin.pm
@@ -8,9 +8,11 @@ use JSON::XS qw( decode_json );
 use Data::Dumper;
 use REST::Client;
 use Storable qw(lock_store lock_retrieve);
+use UUID;
 
 use LINBIT::Linstor;
-use LINBIT::PluginHelper;
+use LINBIT::PluginHelper
+  qw(valid_legacy_name valid_uuid_name valid_snap_name valid_name get_images);
 
 use PVE::Tools qw(run_command trim);
 use PVE::INotify;
@@ -219,8 +221,45 @@ sub volname_and_snap_to_snapname {
     return "snap_${volname}_${snap}";
 }
 
+
+sub uuid_strip_vmid {
+    my ($volname) = @_;
+
+    die "Not a valid uuid volume name ('$volname')"
+      if !valid_uuid_name($volname);
+
+    return substr( $volname, 0, 3 + 8 );
+}
+
+sub pm_name_to_linstor_name {
+    my ($volname) = @_;
+
+    # here we expect contexts that only have lecacy/uuid names, but not snapshot context
+    die "Not a valid volume name ('$volname')" if !valid_name($volname);
+
+    if ( valid_uuid_name($volname) ) {
+        return uuid_strip_vmid($volname);
+    }
+    elsif ( valid_legacy_name($volname) ) {
+        return $volname;
+    }
+    else {
+        die "pm_name_to_linstor_name: '$volname' not valid";
+    }
+}
+
 sub get_dev_path {
-    return "/dev/drbd/by-res/$_[0]/0";
+    my ($volname) = @_;
+
+    # we have to be a bit careful here, this one is called from contexts where the volname can be a snapname
+    die "Not a valid volume name ('$volname')"
+      if !valid_name($volname)
+      and !valid_snap_name($volname);
+
+    # snapshots and legacy names already have their final name
+    $volname = uuid_strip_vmid($volname) if valid_uuid_name($volname);
+
+    return "/dev/drbd/by-res/$volname/0";
 }
 
 
@@ -230,9 +269,11 @@ sub get_dev_path {
 sub map_volume {
     my ( $class, $storeid, $scfg, $volname, $snap ) = @_;
 
-    $volname = volname_and_snap_to_snapname( $volname, $snap ) if defined($snap);
+    my $linstor_name = pm_name_to_linstor_name($volname);
+    $volname = volname_and_snap_to_snapname( $linstor_name, $snap )
+      if defined($snap);
 
-    return get_dev_path "$volname";
+    return get_dev_path($volname);
 }
 
 # For APIVER 2
@@ -245,9 +286,11 @@ sub parse_volname {
 
     if ( $volname =~ m/^(vm-(\d+)-[a-z][a-z0-9\-\_\.]*[a-z0-9]+)$/ ) {
         return ( 'images', $1, $2, undef, undef, undef, 'raw' );
+    } elsif ( $volname =~ m/^(pm-[\da-f]{8}_(\d+))$/ ) {
+        return ( 'images', $1, $2, undef, undef, undef, 'raw' );
     }
 
-    die "unable to parse lvm volume name '$volname'\n";
+    die "unable to parse PVE volume name '$volname'\n";
 }
 
 sub filesystem_path {
@@ -257,7 +300,7 @@ sub filesystem_path {
 
     my ( $vtype, $name, $vmid ) = $class->parse_volname($volname);
 
-    my $path = get_dev_path "$volname";
+    my $path = get_dev_path($volname);
 
     return wantarray ? ( $path, $vmid, $vtype ) : $path;
 }
@@ -282,40 +325,56 @@ sub alloc_image {
 
     die "unsupported format '$fmt'" if $fmt ne 'raw';
 
-    die "illegal name '$name' - should be 'vm-$vmid-*'\n"
-      if defined($name) && $name !~ m/^vm-$vmid-/;
+    my $linstor_name;
+    my $proxmox_name;
 
     my $lsc       = linstor($scfg);
     my $resources = $lsc->get_resources();
 
-    die "volume '$name' already exists\n"
-      if defined($name) && exists $resources->{$name};
+    # pvesm defines that '' is used to flag that a name should be generated
+    # so we don't use 'defined()', but a plain 'if':
+    if ($name) {
+        if (valid_uuid_name($name)) {
+            $proxmox_name = $name;
+            $linstor_name = uuid_strip_vmid($proxmox_name);
+        } elsif (valid_legacy_name($name)) {
+            $proxmox_name = $name;
+            $linstor_name = $proxmox_name;
+        }
+        else {
+            die "allocated name ('$name') has to be a valid UUID or legacy name";
+        }
 
-    if ( !defined($name) ) {
+        die "volume '$linstor_name' already exists\n"
+          if exists $resources->{$linstor_name};
+    }
+    else {
         for ( my $i = 1 ; $i < 100 ; $i++ ) {
-            my $tn = "vm-$vmid-disk-$i";
+            my $tn = "pm-" . lc substr( UUID::uuid(), 0, 8 );
             if ( !exists( $resources->{$tn} ) ) {
-                $name = $tn;
+                $linstor_name = $tn;
+                $proxmox_name = "${tn}_${vmid}";
                 last;
             }
         }
     }
 
     die "unable to allocate an image name for VM $vmid in storage '$storeid'\n"
-      if !defined($name);
+      if !defined($proxmox_name) or !defined($linstor_name);
 
     eval {
         my $res_grp         = get_resource_group($scfg);
         my $local_node_name = get_preferred_local_node($scfg);
         if ( defined($local_node_name) ) {
             print "\nNOTICE\n"
-              . "  Trying to create diskful resource ($name) on ($local_node_name).\n";
+              . "  Trying to create diskful resource ($linstor_name) on ($local_node_name).\n";
         }
-        $lsc->create_resource( $name, $size, $res_grp, $local_node_name );
+        $lsc->create_resource( $linstor_name, $size, $res_grp, $local_node_name );
+        $lsc->set_vmid( $linstor_name, $vmid ); # does not hurt, even vor legacy names
     };
     confess $@ if $@;
 
-    return $name;
+    return $proxmox_name;
 }
 
 sub free_image {
@@ -324,17 +383,19 @@ sub free_image {
     my $lsc = linstor($scfg);
     my $in_use = 1;
 
+    my $linstor_name = pm_name_to_linstor_name($volname);
+
     foreach (0..9) {
       my $resources = $lsc->update_resources();
-      $in_use = $resources->{$volname}->{in_use};
+      $in_use = $resources->{$linstor_name}->{in_use};
       last if (! $in_use);
       sleep(1);
     }
 
-    warn "Resource $volname still in use after giving it some time" if ($in_use);
+    warn "Resource $linstor_name still in use after giving it some time" if ($in_use);
 
     # yolo, what else should we do...
-    eval { $lsc->delete_resource($volname); };
+    eval { $lsc->delete_resource($linstor_name); };
     confess $@ if $@;
 
     return undef;
@@ -348,7 +409,8 @@ sub list_images {
     # unlike 'status' this is not called in loops like crazy, we can get a current view via update_...():
     $cache->{"linstor:resources"} = linstor($scfg)->update_resources()
       unless $cache->{"linstor:resources"};
-    $cache->{"linstor:resource_definitions"} = linstor($scfg)->update_resource_definitions()
+    $cache->{"linstor:resource_definitions"} =
+      linstor($scfg)->update_resource_definitions()
       unless $cache->{"linstor:resource_definitions"};
 
     # TODO:
@@ -357,11 +419,11 @@ sub list_images {
     # be useful to have all vm images in one "consistency group".
     # On the other hand this would make moving disks hard(er).
 
-    my $resources = $cache->{"linstor:resources"};
+    my $resources            = $cache->{"linstor:resources"};
     my $resource_definitions = $cache->{"linstor:resource_definitions"};
-    my $res_grp = get_resource_group($scfg);
+    my $res_grp              = get_resource_group($scfg);
 
-    return LINBIT::PluginHelper::get_images( $storeid, $vmid, $vollist,
+    return get_images( $storeid, $vmid, $vollist,
         $resources, $nodename, $res_grp, $resource_definitions );
 }
 
@@ -417,25 +479,27 @@ sub deactivate_storage {
 sub activate_volume {
     my ( $class, $storeid, $scfg, $volname, $snap, $cache ) = @_;
 
+    my $linstor_name = pm_name_to_linstor_name($volname);
+
     if ($snap) {    # need to create this resource from snapshot
-        my $snapname = volname_and_snap_to_snapname( $volname, $snap );
+        my $snapname = volname_and_snap_to_snapname( $linstor_name, $snap );
         my $new_volname = $snapname;
         my $lsc = linstor($scfg);
         if ( !$lsc->resource_exists($new_volname) ) {
             eval {
-                $lsc->restore_snapshot( $volname, $snapname, $new_volname );
+                $lsc->restore_snapshot( $linstor_name, $snapname, $new_volname );
             };
             confess $@ if $@;
         }
-        $volname = $new_volname; # for the rest of this function switch the name
+        $linstor_name = $new_volname; # for the rest of this function switch the name
     }
 
     my $nodename = PVE::INotify::nodename();
 
-    eval { linstor($scfg)->activate_resource( $volname, $nodename ); };
+    eval { linstor($scfg)->activate_resource( $linstor_name, $nodename ); };
     confess $@ if $@;
 
-    system ('blockdev --setrw ' . get_dev_path $volname);
+    system ('blockdev --setrw ' . get_dev_path($volname));
 
     return undef;
 }
@@ -445,20 +509,21 @@ sub deactivate_volume {
 
     die "deactivate_volume: snapshot not implemented ($snapname)\n" if $snapname;
 
-    my $nodename = PVE::INotify::nodename();
+    my $nodename     = PVE::INotify::nodename();
+    my $linstor_name = pm_name_to_linstor_name($volname);
 
 # deactivate_resource only removes the assignment if diskless, so this could be a single call.
 # We do all this unnecessary dance to print the NOTICE.
     my $lsc = linstor($scfg);
     my $was_diskless_client =
-      $lsc->resource_exists_intentionally_diskless( $volname, $nodename );
+      $lsc->resource_exists_intentionally_diskless( $linstor_name, $nodename );
 
     if ($was_diskless_client) {
         print "\nNOTICE\n"
-          . "  Intentionally removing diskless assignment ($volname) on ($nodename).\n"
+          . "  Intentionally removing diskless assignment ($linstor_name) on ($nodename).\n"
           . "  It will be re-created when the resource is actually used on this node.\n";
 
-        eval { $lsc->deactivate_resource( $volname, $nodename ); };
+        eval { $lsc->deactivate_resource( $linstor_name, $nodename ); };
         confess $@ if $@;
     }
 
@@ -468,9 +533,10 @@ sub deactivate_volume {
 sub volume_resize {
     my ( $class, $scfg, $storeid, $volname, $size, $running ) = @_;
 
-    my $size_kib = ( $size / 1024 );
+    my $size_kib     = ( $size / 1024 );
+    my $linstor_name = pm_name_to_linstor_name($volname);
 
-    eval { linstor($scfg)->resize_resource( $volname, $size_kib ); };
+    eval { linstor($scfg)->resize_resource( $linstor_name, $size_kib ); };
     confess $@ if $@;
 
     return 1;
@@ -478,9 +544,11 @@ sub volume_resize {
 
 sub volume_snapshot {
     my ( $class, $scfg, $storeid, $volname, $snap ) = @_;
-    my $snapname = volname_and_snap_to_snapname( $volname, $snap );
 
-    eval { linstor($scfg)->create_snapshot( $volname, $snapname ); };
+    my $linstor_name = pm_name_to_linstor_name($volname);
+    my $snapname     = volname_and_snap_to_snapname( $linstor_name, $snap );
+
+    eval { linstor($scfg)->create_snapshot( $linstor_name, $snapname ); };
     confess $@ if $@;
 
     return 1;
@@ -489,9 +557,10 @@ sub volume_snapshot {
 sub volume_snapshot_rollback {
     my ( $class, $scfg, $storeid, $volname, $snap ) = @_;
 
-    my $snapname = volname_and_snap_to_snapname( $volname, $snap );
+    my $linstor_name = pm_name_to_linstor_name($volname);
+    my $snapname     = volname_and_snap_to_snapname( $linstor_name, $snap );
 
-    eval { linstor($scfg)->rollback_snapshot( $volname, $snapname ); };
+    eval { linstor($scfg)->rollback_snapshot( $linstor_name, $snapname ); };
     confess $@ if $@;
 
     return 1;
@@ -499,7 +568,9 @@ sub volume_snapshot_rollback {
 
 sub volume_snapshot_delete {
     my ( $class, $scfg, $storeid, $volname, $snap ) = @_;
-    my $snapname = volname_and_snap_to_snapname( $volname, $snap );
+
+    my $linstor_name = pm_name_to_linstor_name($volname);
+    my $snapname     = volname_and_snap_to_snapname( $linstor_name, $snap );
 
     my $lsc = linstor($scfg);
 
@@ -512,10 +583,31 @@ sub volume_snapshot_delete {
     eval { $lsc->delete_resource( $snapname ); };
     confess $@ if $@;
 
-    eval { $lsc->delete_snapshot( $volname, $snapname ); };
+    eval { $lsc->delete_snapshot( $linstor_name, $snapname ); };
     confess $@ if $@;
 
     return 1;
+}
+
+sub rename_volume {
+    my ( $class, $scfg, $storeid, $source_volname, $target_vmid,
+        $target_volname )
+      = @_;
+
+    die "rename_volume only possible for uuid disk names"
+      if !valid_uuid_name($source_volname);
+
+    die "rename_volume only possible if target name is unset"
+      if defined($target_volname);
+
+    # we keep the UUID name, we just switch to a new VM
+    my $linstor_name = pm_name_to_linstor_name($source_volname);
+    $target_volname = $linstor_name . "_" . $target_vmid;
+
+    eval { linstor($scfg)->set_vmid( $linstor_name, $target_vmid ); };
+    confess $@ if $@;
+
+    return "${storeid}:${target_volname}";
 }
 
 sub volume_has_feature {
@@ -526,6 +618,9 @@ sub volume_has_feature {
         copy     => { base    => 1, current => 1 },
         snapshot => { current => 1 },
     };
+
+    # pretend we can rename stuff, but fail for legacy names.
+    $features->{rename} = {current => 1};
 
     my ( $vtype, $name, $vmid, $basename, $basevmid, $isBase ) =
       $class->parse_volname($volname);