+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Changed
 - include diskless assignments in the GUI's "VM Disks"
+- allow renaming of volumes
 
 ## [7.0.1] - 2023-06-15
 ### Fixed