+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## 8.0.0 - 2024-04-17
 ### Changed
 - include diskless assignments in the GUI's "VM Disks"
 - set 'allow-two-primaries' in RD creation (requires LINSTOR >= 1.27.0)